(refactor) LineCount 사용하지 않게 수정 x축 개수는 user count로 판단
