+++ src/main/java/step02/dto/LadderGameDTO.java
@@ -19,4 +19,8 @@ public class LadderGameDTO {
     public Users getUsers() {
         return users;
     }
+
+    public int getUserCount() {
+        return users.getUserCount();
+    }
 }
