+++ src/main/java/step02/ladder/Ladder.java
@@ -1,28 +1,32 @@
 package step02.ladder;
 
+import step02.exception.ZeroHeightException;
+
 import java.util.ArrayList;
 import java.util.List;
 
 public class Ladder {
-    private final List<Line> lines = new ArrayList<>();
-    private final LineCount yCount;
-    private final LineCount xCount;
+    private final static String ERROR_ZERO_LINE = "사다리는 높이 0이상으로 입력해주세요.";
+    private final static int ZERO = 0;
 
-    public Ladder(int xCount, int yCount, LineStrategy lineStrategy) throws IllegalAccessException {
-        this.yCount = new LineCount(yCount);
-        this.xCount = new LineCount(xCount);
+    private final List<Line> lines = new ArrayList<>();
+
+    public Ladder(int xCount, int yCount, LineStrategy lineStrategy) throws ZeroHeightException {
+        checkZeroLine(yCount);
 
         for (int i = 0; i < yCount; i++) {
             lines.add(new Line(xCount, lineStrategy));
         }
     }
 
-    public int getYLineCount() {
-        return yCount.getCount();
+    private void checkZeroLine(int height) throws ZeroHeightException {
+        if (height <= ZERO) {
+            throw new ZeroHeightException(ERROR_ZERO_LINE);
+        }
     }
 
-    public int getXLineCount() {
-        return xCount.getCount();
+    public int getYLineCount() {
+        return lines.size();
     }
 
     public List<Line> getLadderLine() {
